% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_palettes.R
\name{pt_bwr}
\alias{pt_bwr}
\title{Blue-white-red diverging colour ramp}
\usage{
pt_bwr(n = 256L)
}
\arguments{
\item{n}{number of colours to generate.}
}
\value{
A character vector of \code{n} hex colours.
}
\description{
The R equivalent of the Python \code{R_BWR} colormap: linear RGB interpolation
between blue, white and red.
}
