% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{PT_BWR_COLORS}
\alias{PT_BWR_COLORS}
\title{Blue-white-red colour ramp anchors}
\format{
An object of class \code{character} of length 3.
}
\usage{
PT_BWR_COLORS
}
\description{
Matches \code{colorRamp2(c(-2, 0, 2), c("blue", "white", "red"))}.
}
\keyword{datasets}
