% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{pt_gene_universe}
\alias{pt_gene_universe}
\title{Union of all genes across every gene set}
\usage{
pt_gene_universe(gs)
}
\arguments{
\item{gs}{a named list of gene sets, e.g. from \code{\link[=pt_parse_gmt]{pt_parse_gmt()}}.}
}
\value{
A character vector of unique gene identifiers.
}
\description{
Union of all genes across every gene set
}
