% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_stats.R
\name{pt_run_test}
\alias{pt_run_test}
\title{Dispatch a two-group test by name}
\usage{
pt_run_test(method, case, ref)
}
\arguments{
\item{method}{one of \code{"welch"}, \code{"mannwhitney"}, \code{"moderated_t"}.}

\item{case, ref}{pathway x sample matrices for the two groups.}
}
\value{
A list of \code{effect}, \code{statistic}, \code{p_value}.
}
\description{
Dispatch a two-group test by name
}
