% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{pt_samples}
\alias{pt_samples}
\alias{pt_features}
\alias{pt_terms}
\title{Row and column labels of a matrix contract}
\usage{
pt_samples(x)

pt_features(x)

pt_terms(x)
}
\arguments{
\item{x}{a \code{\link[=pt_feature_matrix]{pt_feature_matrix()}} or \code{\link[=pt_score_matrix]{pt_score_matrix()}}.}
}
\value{
A character vector.
}
\description{
Row and column labels of a matrix contract
}
