% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\docType{data}
\name{PT_SCORED}
\alias{PT_SCORED}
\alias{PT_NO_OVERLAP}
\alias{PT_TOO_SMALL}
\alias{PT_TOO_LARGE}
\title{Fates a requested gene set can meet}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
PT_SCORED

PT_NO_OVERLAP

PT_TOO_SMALL

PT_TOO_LARGE
}
\description{
Fates a requested gene set can meet
}
\keyword{datasets}
