% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{pt_silhouette}
\alias{pt_silhouette}
\title{Mean silhouette width of a clustering}
\usage{
pt_silhouette(x, labels)
}
\arguments{
\item{x}{an observations x features numeric matrix.}

\item{labels}{cluster assignment per observation.}
}
\value{
The mean silhouette width, or \code{NA} when fewer than two clusters.
}
\description{
Mean silhouette width of a clustering
}
