% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{pt_subset_samples}
\alias{pt_subset_samples}
\title{Restrict a score matrix to a set of samples}
\usage{
pt_subset_samples(sm, columns)
}
\arguments{
\item{sm}{a \code{\link[=pt_score_matrix]{pt_score_matrix()}}.}

\item{columns}{sample ids to keep, in the order wanted.}
}
\value{
A \code{pt_score_matrix}.
}
\description{
Restrict a score matrix to a set of samples
}
