% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_stats.R
\name{pt_trigamma_inverse}
\alias{pt_trigamma_inverse}
\title{Solve \code{trigamma(y) = x} for \code{y}}
\usage{
pt_trigamma_inverse(x)
}
\arguments{
\item{x}{a numeric scalar.}
}
\value{
The solution \code{y}.
}
\description{
A port of \code{limma::trigammaInverse}.
}
